Working implementation of the ExtensionPoint engine
