+++ src/main/resources/org/jenkinsci/plugins/securityinspector/Messages.properties
@@ -1,3 +1,3 @@
-JobReport.RowColumnHeader=Jobs
-SlaveReport.RowColumnHeader=Slaves
+JobReport.RowColumnHeader=Items
+SlaveReport.RowColumnHeader=Computers
 UserReport.RowColumnHeader=Users
