+++ src/main/resources/org/jenkinsci/plugins/securityinspector/SecurityInspectorAction/index.jelly
@@ -27,14 +27,13 @@
          xmlns:t="/lib/hudson" xmlns:f="/lib/form">
     <l:layout permission="${app.ADMINISTER}" xmlns:local="local" norefresh="true">
         <st:include it="${app}" page="sidepanel.jelly"/>
-
+        
         <d:taglib uri="local">
             <!-- table to show a map -->
-            <d:tag name="feature">
-                <t:summary icon="${icon}"
-                       href="${href}" iconOnly="true">
+            <d:tag name="reportBuilder">
+                <t:summary icon="${icon}" href="${href}" iconOnly="true">
                     <div class="link">
-                        <a href="${href}">${title}</a>
+                        <a href="${href}/createReport">${title}</a>
                     </div>
                     <div style="color:gray; text-decoration:none;">
                         <d:invokeBody />
@@ -46,13 +45,18 @@
         <d:taglib uri="local">
             <!-- table to show a map -->
             <d:tag name="category">
-              <h2>${title}</h2>
+              
               <j:set var="reportBuiders" value="${it.getReportBuilders(type)}"/>
-              <j:if test="${reportBuiders.isEmpty()}">
-                ${%No available reports for this type}
-              </j:if>
-              <j:if test="${reportBuiders.size() != 0}">
-                ${%Reports available: } ${reportBuiders.size()}
+              <j:if test="${!reportBuiders.isEmpty()}">
+                <table style="padding-left: 2em;" id="management-links">
+                  <h2>${title}</h2>
+                  <j:forEach var="reportBuilder" items="${reportBuiders}">
+                    <local:reportBuilder icon="${reportBuilder.icon}" href="${reportBuilder.index}" 
+                                   title="${reportBuilder.displayName}">
+                        ${reportBuilder.description}
+                    </local:reportBuilder>
+                  </j:forEach>
+                </table>
               </j:if>
             </d:tag>
         </d:taglib>
@@ -62,35 +66,9 @@
                 <img src="${imagesURL}/48x48/${it.iconFileName}" alt="${%Security Inspector}"/>
                 ${%Security Inspector}
             </h1>
-
             <local:category title="${%Users}" type="USER"/>
             <local:category title="${%Jobs}" type="ITEM"/>
             <local:category title="${%Computers}" type="COMPUTER"/>
-                
-            <h1>
-              Old UI
-            </h1>
-
-
-            <h2>${%Users}</h2>
-            <table style="padding-left: 2em;" id="management-links">
-                <local:feature icon="fingerprint.gif" href="job-filter" 
-                               title="${%Single user, multiple jobs}">
-                    ${%Display job permissions for the specified user}
-                </local:feature>
-                <local:feature icon="fingerprint.gif" href="slave-filter" 
-                               title="${%Single user, multiple nodes}">
-                    ${%Display node permissions for the specified user}
-                </local:feature>                        
-            </table>
-            
-            <h2>${%Jobs}</h2>
-            <table style="padding-left: 2em;" id="management-links">
-                  <local:feature icon="user.png" href="user-filter" 
-                               title="${%Multiple users, single job}">
-                    ${%Display users permissions for the specified job}
-                </local:feature> 
-            </table>
         </l:main-panel>
     </l:layout>
 </j:jelly>
