--- src/main/resources/org/jenkinsci/plugins/securityinspector/SecurityInspectorAction/search_report_user_4_slave.jelly
@@ -1,109 +1,0 @@
-<!--
- * The MIT License
- *
- * Copyright 2014 Ksenia Nenasheva <ks.nenasheva@gmail.com>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- -->
- <?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-    <l:layout title="${%Security Inspector}" permission="${app.ADMINISTER}">
-        <st:include it="${app}" page="sidepanel.jelly"/>
-        <l:main-panel>
-            <link rel="stylesheet" href="${rootURL}${app.VIEW_RESOURCE_PATH}/hudson/security/table.css" type="text/css" />
-            <link rel="stylesheet" href="${rootURL}/plugin/security-inspector/css/security-inspector.css" type="text/css" />    
-             
-            <j:set var="_user" value="${it.requestedUser}"/>
-      
-            <f:form method="post" name="config" action="goHome">
-                <h1>${%Security Inspector. Report for user} ${_user.id}</h1> 
-                
-                <f:block>
-                                   
-                    <!--Slave -->
-                    <h2>${%Slave reports}</h2>
-                    <j:set var="_report" value="${it.reportSlave}"/>
-                    <table id="projectPermissions" name="data">
-                      
-                      <j:if test="${_report.groups.size() == null}">
-                        <h3>${%Report is empty}</h3>
-                      </j:if>
-                      
-                      <j:if test="${_report.groups.size() != null}">
-
-                        <!-- The first row will show grouping -->
-                        <tr class="group-row">
-                            <td rowspan="2" class="start" />
-                            <td rowspan="2" class="pane-header blank">
-                                ${_report.rowColumnHeader}
-                            </td>
-                                                
-                            <j:forEach var="group" items="${_report.groups}">                          
-                                <td class="pane-header" colspan="${_report.getItemsOfGroup(group).size()}">
-                                    ${_report.getGroupTitle(group)}
-                                </td>
-                            </j:forEach>
-                        
-                            <td rowspan="2" class="stop" />
-                        </tr> 
-                        
-                        <!-- Second Row -->     
-                        <tr class="caption-row">
-                            <j:forEach var="group" items="${_report.groups}">
-                                <j:forEach var="groupItem" items="${_report.getItemsOfGroup(group)}">
-                                    <td class="pane">
-                                        ${_report.getColumnTitle(groupItem)}
-                                    </td>
-                                </j:forEach>
-                            </j:forEach>
-                        </tr>
-                            
-                        <!-- Body -->
-                        <j:forEach var="row" items = "${_report.rows}">                       
-                            <tr>
-                                <td class="start" />
-                                <td style="white-space: nowrap">${_report.getRowTitle(row)}</td>
-                                <j:forEach var="group" items="${_report.groups}">
-                                    <j:forEach var="groupItem" items="${_report.getItemsOfGroup(group)}">
-                                        <td class="pane">
-                                            <j:set var="_entryReport" value="${_report.entries.get(row,groupItem)}"/>
-                                            <j:if test="${_report.isEntryReportOk(row,groupItem,_entryReport) == true}">
-                                                &#10004;
-                                            </j:if>
-                                        </td>
-                                    </j:forEach>
-                                </j:forEach>
-                                <td class="stop" />
-                            </tr>                       
-                        </j:forEach>
-                      </j:if>
-     
-                    </table>
-                </f:block>
-                <br />
-                <br /> 
-                <f:block>
-                                    
-                    <f:submit name="GoToSF" value="${%Back}" />
-                </f:block>
-                
-            </f:form>
-        </l:main-panel>
-    </l:layout>
-</j:jelly>
