--- src/main/resources/org/jenkinsci/plugins/securityinspector/SecurityInspectorAction/slave-filter.jelly
@@ -1,75 +1,0 @@
-<!--
-* The MIT License
-*
-* Copyright 2014 Ksenia Nenasheva <ks.nenasheva@gmail.com>
-*
-* Permission is hereby granted, free of charge, to any person obtaining a copy
-* of this software and associated documentation files (the "Software"), to deal
-* in the Software without restriction, including without limitation the rights
-* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-* copies of the Software, and to permit persons to whom the Software is
-* furnished to do so, subject to the following conditions:
-*
-* The above copyright notice and this permission notice shall be included in
-* all copies or substantial portions of the Software.
-*
-* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-* THE SOFTWARE.
--->
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-    <l:layout title="${%Security Inspector}" permission="${app.ADMINISTER}">
-        <st:include it="${app}" page="sidepanel.jelly"/>
-        <l:main-panel>
-            
-            <!-- Job Report -->
-             <f:form method="post" name="config" action="filterSubmit">
-                <h1>
-                    <img src="${imagesURL}/48x48/${it.iconFileName}" alt="${%Security Inspector}"/>
-                    ${%Security Inspector}
-                </h1>
-                      
-                <f:section title="${%Report for user}"> 
-                    <f:entry title="${%Selected user:}">
-                        <j:set var="_userList"  value="${it.helper.possibleUsers}"/>
-                        <select name="selectedUser" class="setting-input">
-                            <j:forEach var="user" items="${_userList}">
-                                <f:option value="${user.getId()}">
-                                    ${it.helper.getDisplayName(user)}
-                                </f:option>
-                            </j:forEach>
-                        </select> 
-                    </f:entry>
-                    
-                    <f:block> 
-                        <f:entry>
-                            <table width="100%"> 
-                                <f:section title="${%Filters slaves to be reported}">                               
-                                                             
-                                    <f:optionalBlock name="useincluderegex4slave" title="${%Use a regular expression to include slaves into the view}"
-                                                     checked="${it.filters.includeRegex4Slave != null}" >
-                                        <f:entry title="${%Regular expression for slave}" field="includeRegex4Slave">
-                                            <f:textbox value="${it.filters.includeRegex4Slave}"
-                                                        checkUrl="'${rootURL}/plugin/security-inspector/checkRegex?regex='+this.value"/>
-                                        </f:entry>
-                                    </f:optionalBlock>
-                                </f:section>
-                            </table>
-                        </f:entry>
-                    </f:block>
-                    
-                    <f:block>
-                        <f:submit name="Submit4slaves" value="${%Submit}" />
-                        <f:submit name="GoToHP" value="${%Back}" />
-                    </f:block>
-
-                </f:section>
-            </f:form>
-        </l:main-panel>
-    </l:layout>
-</j:jelly>
