+++ src/main/resources/org/jenkinsci/plugins/securityinspector/impl/items/ItemForMultipleUsersReportBuilder/configureReport.jelly
@@ -1,0 +1,54 @@
+<!--
+* The MIT License
+*
+* Copyright 2014 Ksenia Nenasheva <ks.nenasheva@gmail.com>
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in
+* all copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+* THE SOFTWARE.
+-->
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+  <j:set var="_jobList" value="${it.possibleJobs}"/>      
+  <f:entry title="${%Selected job: }"> 
+    <select name="selectedJobs" class="setting-input">
+      <j:forEach var="job" items="${_jobList}">
+        <f:option value="${job.fullName}">
+          ${job.fullDisplayName}
+        </f:option>
+      </j:forEach>
+    </select>                        
+  </f:entry>
+                    
+  <f:block> 
+    <f:entry>
+      <table width="100%"> 
+        <f:section title="${%Filter users to be reported}"> 
+                    
+          <f:optionalBlock name="useincluderegex4user" 
+                                                     title="${%Filter user IDs by a regular expression}"
+                                                     checked="${it.filters.includeRegex4User != null}" >
+            <f:entry title="${%Regular expression}" field="includeRegex4User">
+              <f:textbox value="${it.filters.includeRegex4User}"
+                                                       checkUrl="'${rootURL}/plugin/security-inspector/checkRegex?regex='+this.value"/>
+            </f:entry>
+          </f:optionalBlock>
+        </f:section>
+      </table>
+    </f:entry>
+  </f:block>
+</j:jelly>
