+++ src/main/resources/org/jenkinsci/plugins/securityinspector/impl/users/PermissionsForComputerReportBuilder/reportBody.jelly
@@ -1,0 +1,90 @@
+<!--
+* The MIT License
+*
+* Copyright 2014 Ksenia Nenasheva <ks.nenasheva@gmail.com>
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in
+* all copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+* THE SOFTWARE.
+-->
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form"> 
+  <j:set var="_user" value="${it.requestedUser}"/>
+  <j:set var="_report" value="${it.reportSlave}"/>
+  <h3>${%Permissions of} &quot;${_user.id}&quot; ${%for computers}</h3>
+                
+  <f:block>                  
+    <!-- Computer -->
+    
+    <table id="projectPermissions" name="data">
+                      
+      <j:if test="${_report.groups.size() == null}">
+        <h3>${%Report is empty}</h3>
+      </j:if>
+                      
+      <j:if test="${_report.groups.size() != null}">
+
+        <!-- The first row will show grouping -->
+        <tr class="group-row">
+          <td rowspan="2" class="start" />
+          <td rowspan="2" class="pane-header blank">
+            ${_report.rowColumnHeader}
+          </td>
+                                                
+          <j:forEach var="group" items="${_report.groups}">                          
+            <td class="pane-header" colspan="${_report.getItemsOfGroup(group).size()}">
+              ${_report.getGroupTitle(group)}
+            </td>
+          </j:forEach>
+                        
+          <td rowspan="2" class="stop" />
+        </tr> 
+                        
+        <!-- Second Row -->     
+        <tr class="caption-row">
+          <j:forEach var="group" items="${_report.groups}">
+            <j:forEach var="groupItem" items="${_report.getItemsOfGroup(group)}">
+              <td class="pane">
+                ${_report.getColumnTitle(groupItem)}
+              </td>
+            </j:forEach>
+          </j:forEach>
+        </tr>
+                            
+        <!-- Body -->
+        <j:forEach var="row" items = "${_report.rows}">                       
+          <tr>
+            <td class="start" />
+            <td style="white-space: nowrap">${_report.getRowTitle(row)}</td>
+            <j:forEach var="group" items="${_report.groups}">
+              <j:forEach var="groupItem" items="${_report.getItemsOfGroup(group)}">
+                <td class="pane">
+                  <j:set var="_entryReport" value="${_report.entries.get(row,groupItem)}"/>
+                  <j:if test="${_report.isEntryReportOk(row,groupItem,_entryReport) == true}">
+                      &#10004;
+                  </j:if>
+                </td>
+              </j:forEach>
+            </j:forEach>
+            <td class="stop" />
+          </tr>                       
+        </j:forEach>
+      </j:if>
+     
+    </table>
+  </f:block>
+</j:jelly>
