+++ src/main/resources/org/jenkinsci/plugins/securityinspector/impl/users/PermissionsForItemReportBuilder/configureReport.jelly
@@ -1,0 +1,38 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+  <f:entry title="${%Selected user:}">
+    <j:set var="_userList"  value="${it.possibleUsers}"/>
+    <select name="selectedUser" class="setting-input">
+      <j:forEach var="user" items="${_userList}">
+        <f:option value="${user.getId()}">
+          ${it.getDisplayName(user)}
+        </f:option>
+      </j:forEach>
+    </select> 
+  </f:entry>
+                        
+  <f:block> 
+    <f:entry>
+      <table width="100%"> 
+        <f:section title="${%Filter jobs to be reported}">                                                         
+          <f:optionalBlock name="useincluderegex" title="${%Use a regular expression to include jobs into the view}"
+                                                     checked="${it.filters.includeRegex != null}" >
+            <f:entry title="${%Regular expression for jobs}" field="includeRegex">
+              <f:textbox value="${it.filters.includeRegex}"
+                                                       checkUrl="'${rootURL}/plugin/security-inspector/checkRegex?regex='+this.value"/>
+            </f:entry>
+          </f:optionalBlock>
+          <j:invokeStatic var="allJobFilters" className="hudson.views.ViewJobFilter" method="all"/>                   
+          <j:if test="${allJobFilters.size() != 0}">
+            <f:block>
+              <f:hetero-list name="jobFilters" hasHeader="true"
+                                                           descriptors="${allJobFilters}"
+                                                           addCaption="${%Add Job Filter}" />
+            </f:block>
+          </j:if>
+                                    
+        </f:section>
+      </table>
+    </f:entry>
+  </f:block>
+</j:jelly>
