+++ src/main/resources/org/jenkinsci/plugins/securityinspector/model/ReportBuilder/report.jelly
@@ -1,0 +1,25 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+    <l:layout title="${%Security Inspector}" permission="${app.ADMINISTER}">
+        <st:include it="${app}" page="sidepanel.jelly"/>
+        <l:main-panel>
+            <link rel="stylesheet" href="${rootURL}${app.VIEW_RESOURCE_PATH}/hudson/security/table.css" type="text/css" />
+            <link rel="stylesheet" href="${rootURL}/plugin/security-inspector/css/security-inspector.css" type="text/css" />   
+            
+            <h1>
+              <img src="${imagesURL}/48x48/secure.png" alt="${%Security Inspector}"/>
+              ${%Security Inspector}
+            </h1>
+             
+            <!-- Report with a table and the return button -->     
+            <f:form method="post" name="config" action="goHome">
+                <st:include page="reportBody.jelly" optional="false"/>
+                <br/>
+                <br/> 
+                <f:block>                 
+                    <f:submit name="goHomeButton" value="${%Back}" />
+                </f:block>
+            </f:form>
+        </l:main-panel>
+    </l:layout>
+</j:jelly>
